+++ packages/terra-date-picker/src/react-datepicker/week.jsx
@@ -102,10 +102,6 @@ export default class Week extends React.Component {
      * Difference between utc and local time.
      */
     utcOffset: PropTypes.number,
-    /**
-     * Whether or not to apply keyboard selected styles.
-     */
-    disableKeyboardSelected: PropTypes.bool,
   }
 
   handleDayClick = (day, event) => {
@@ -167,8 +163,7 @@ export default class Week extends React.Component {
           startDate={this.props.startDate}
           endDate={this.props.endDate}
           dayClassName={this.props.dayClassName}
-          utcOffset={this.props.utcOffset}
-          disableKeyboardSelected={this.props.disableKeyboardSelected}/>
+          utcOffset={this.props.utcOffset}/>
       )
     }))
   }